Refactor: Conditionally hide menu in AdventureItem
This commit introduces a `showMenu` parameter to the `AdventureItem` composable. When `showMenu` is false, the three-dot menu icon and its associated `DropdownMenu` will not be displayed.

This change is used in `HomeContent.kt` to hide the menu for adventures displayed on the home screen.

The internal state variable for the dropdown menu visibility has been renamed from `showMenu` to `showDropdownMenu` to avoid confusion with the new parameter.